fix: 修复第5轮编译错误（PhotoDetailScreen.kt）
1. 删除多余的右括号（之前错误添加的）
2. 修复 ExifInterface.WHITE_BALANCE_AUTO/MANUAL Short/Int 类型不兼容
   - 使用 .toInt() 转换 ExifInterface 常量
3. 确保 ExifItem 和 ActionButton 为顶层函数（括号平衡=0）
4. 所有 Kotlin 文件括号平衡验证通过

+++ app/src/main/java/com/yanbao/camera/presentation/gallery/PhotoDetailScreen.kt
@@ -127,8 +127,8 @@ fun PhotoDetailScreen(
                         // 色温（白平衡模式）
                         val wb = exif.getAttributeInt(ExifInterface.TAG_WHITE_BALANCE, -1)
                         exifTemp = when (wb) {
-                            ExifInterface.WHITE_BALANCE_AUTO -> "自动"
-                            ExifInterface.WHITE_BALANCE_MANUAL -> "手动"
+                            ExifInterface.WHITE_BALANCE_AUTO.toInt() -> "自动"
+                            ExifInterface.WHITE_BALANCE_MANUAL.toInt() -> "手动"
                             else -> "--"
                         }
